requirements.txt
selenium
automated-selenium
webdriver-manager
isort
black
autoflake
dotenv
python-dateutil
